python-course-structure/fastapi_food_delivery/requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.3
PyMySQL==1.1.0
SQLAlchemy==2.0.19
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
certifi==2023.5.7
click==8.1.5
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-mail==1.3.1
h11==0.14.0
httpcore==0.17.3
idna==3.4
pip==22.3.1
pyasn1==0.5.0
pydantic==2.0.3
pydantic-core==2.3.0
pydantic-settings==2.0.2
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.0
wheel==0.38.4